+++ NEWS.md
@@ -6,6 +6,9 @@
 * New `httpbin_app()` end point: `/stream-bytes`, to send chunked
   random bytes.
 
+* The web server now does not crash if the connection is closed by the
+  client while delaying the response.
+
 # 1.0.0
 
 First public release.